IDE-2774 move to IOUtil
