+++ tools/plugins/com.liferay.ide.project.ui/src/com/liferay/ide/project/ui/upgrade/action/ImportProjectNextActionHandler.java
@@ -28,17 +28,15 @@ import org.eclipse.core.runtime.NullProgressMonitor;
 import org.eclipse.core.runtime.Path;
 import org.eclipse.core.runtime.Platform;
 import org.eclipse.sapphire.platform.PathBridge;
-import org.eclipse.sapphire.platform.StatusBridge;
 import org.eclipse.sapphire.ui.Presentation;
 
 import com.liferay.ide.core.ILiferayProjectImporter;
 import com.liferay.ide.core.LiferayCore;
-import com.liferay.ide.core.util.FileUtil;
+import com.liferay.ide.core.util.IOUtil;
 import com.liferay.ide.core.util.ZipUtil;
 import com.liferay.ide.project.core.ProjectCore;
 import com.liferay.ide.project.core.modules.BladeCLI;
 import com.liferay.ide.project.core.modules.BladeCLIException;
-import com.liferay.ide.project.core.util.ProjectImportUtil;
 import com.liferay.ide.project.core.util.ProjectUtil;
 import com.liferay.ide.project.ui.ProjectUI;
 import com.liferay.ide.sdk.core.SDK;
@@ -131,7 +129,7 @@ public class ImportProjectNextActionHandler extends BaseActionHandler
 
         ZipUtil.unzip( sdkZipFile, stateDir );
 
-        FileUtil.copyDirToDir( new File( stateDir, "com.liferay.portal.plugins.sdk-7.0" ), targetSDKLocation.toFile() );
+        IOUtil.copyDirToDir( new File( stateDir, "com.liferay.portal.plugins.sdk-7.0" ), targetSDKLocation.toFile() );
     }
 
     private void createLiferayWorkspace( IPath targetSDKLocation, IProgressMonitor monitor ) throws BladeCLIException
